pyproject.toml
[project]
name = "workspaces"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "microsoft-kiota-abstractions>=1.9.3",
    "microsoft-kiota-authentication-azure>=1.9.3",
    "microsoft-kiota-bundle>=1.9.3",
    "microsoft-kiota-http>=1.9.3",
    "microsoft-kiota-serialization-json>=1.9.3",
    "microsoft-kiota-serialization-text>=1.9.3",
    "pytest>=8.3.5",
]
